refactor: remove backward compatibility modules and update imports
- Updated all test file imports (18 files) to use new module paths
- Updated all documentation imports (19 files) to use new module paths
- Updated example file imports (2 files) to use new module paths
- Removed backward compatibility references from __init__.py
- Deleted 8 backward compatibility files:
  - exceptions.py
  - function_calls.py
  - multimodal.py
  - process_response.py
  - retry.py
  - schema_utils.py
  - validators.py
  - dsl/validators.py

All tests pass successfully with the new import paths.
