+++ docs/integrations/openai.md
@@ -169,7 +169,7 @@ Let's see an example below with the sample image above where we'll load it in us
 Note that we support local files and base64 strings too with the `from_path` and the `from_base64` class methods.
 
 ```python
-from instructor.multimodal import Image
+from instructor.processing.multimodal import Image
 from pydantic import BaseModel, Field
 import instructor
 from openai import OpenAI
@@ -222,7 +222,7 @@ Let's see an example below with the sample PDF above where we'll load it in usin
 Note that we support local files and base64 strings too with the `from_path` and the `from_base64` class methods.
 
 ```python
-from instructor.multimodal import PDF
+from instructor.processing.multimodal import PDF
 from pydantic import BaseModel, Field
 import instructor
 from openai import OpenAI
@@ -267,7 +267,7 @@ Instructor makes it easy to analyse and extract semantic information from Audio 
 Note that we support local files and base64 strings too with the `from_path`
 
 ```python
-from instructor.multimodal import Audio
+from instructor.processing.multimodal import Audio
 from pydantic import BaseModel
 import instructor
 from openai import OpenAI
