--- instructor/function_calls.py
@@ -1,4 +1,0 @@
-"""Backward compatibility module for instructor.function_calls imports."""
-
-# Re-export everything from the new location
-from .processing.function_calls import *  # noqa: F403, F401
