--- instructor/validators.py
@@ -1,6 +1,0 @@
-"""Backward compatibility module for instructor.validators imports."""
-
-# Re-export everything from the new location
-from .validation.async_validators import *  # noqa: F403, F401
-
-# This allows `from instructor.validators import AsyncValidationContext` to work
