+++ tests/llm/test_openai/test_multimodal.py
@@ -1,5 +1,5 @@
 import pytest
-from instructor.multimodal import Image, Audio
+from instructor.processing.multimodal import Image, Audio
 import instructor
 from pydantic import Field, BaseModel
 from itertools import product
