+++ tests/llm/test_vertexai/test_deprecated_async.py
@@ -2,7 +2,7 @@ import pytest
 from unittest.mock import patch, MagicMock
 from pydantic import BaseModel
 from instructor import from_vertexai
-from instructor.exceptions import ConfigurationError
+from instructor.core.exceptions import ConfigurationError
 
 
 class User(BaseModel):
